Create storyboard segues.
I created a vc segue to PokeMatchVC for the BG button with an alert and
dismiss and continue. Created another at the game over function that
brings up an alert with scores, brings you to the winner screen with
final scores. Need to the add an interval timer to then bring you to
GC. Also, GC Done button not working in Leaderboards VC.